+++ src/presenter/film-popup.js
@@ -1,7 +1,8 @@
 ﻿import FilmPopupView from '../view/film-popup.js';
 import PopupCommentView from '../view/popup-comment.js';
 import PopupControlsView from '../view/popup-controls.js';
-import { render, remove, RenderPosition } from '../utils.js';
+import { render, remove } from '../utils.js';
+import { RenderPosition } from '../consts.js';
 
 export default class FilmPopup {
   constructor(container, changeData, profieName) {
