+++ src/utils.js
@@ -1,34 +1,10 @@
-﻿import Abstract from './view/abstract.js';
+﻿import { SortType } from './consts.js';
+import Abstract from './view/abstract.js';
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
 import relativeTime from 'dayjs/plugin/relativeTime';
 dayjs.extend(duration);
 dayjs.extend(relativeTime);
-
-export const UserAction = {
-  UPDATE_FILM: 'UPDATE_FILM',
-  ADD_COMMENT: 'ADD_COMMENT',
-  DELETE_COMMENT: 'DELETE_COMMENT',
-};
-
-export const UpdateType = {
-  PATCH: 'PATCH',
-  MINOR: 'MINOR',
-  MAJOR: 'MAJOR',
-};
-
-export const SortType = {
-  DEFAULT: 'default',
-  BY_DATE: 'date',
-  BY_RATING: 'rating',
-  BY_COMMENT_AMOUNT: 'comment_amount',
-};
-
-export const RenderPosition = {
-  AFTERBEGIN: 'afterbegin',
-  AFTEREND: 'afterend',
-  BEFOREEND: 'beforeend',
-};
 
 export const SortStrategy = {
   [SortType.BY_RATING]: (a, b) => b.filmInfo.rating - a.filmInfo.rating,
